Move O_LARGEFILE handling into internal/osutil
Address review feedback on the strict-root large-file fix:

- Move the StrictSystemRoot wrapper and the openLargeFile constant from
  pkg/fs into internal/osutil, alongside the other os-specific helpers.
  The constant cannot live in pkg/fs and be referenced from a wrapper
  there while osutil imports pkg/fs (filetimes) without creating an import
  cycle, so the wrapper moves too; it only depends on os.
- Drop the redundant "//go:build linux" tag. Instead of a filename OS
  suffix, use an explicit "linux || solaris || aix || zos" constraint,
  which are exactly the platforms where golang.org/x/sys/unix defines
  O_LARGEFILE. A plain "unix" constraint would fail to compile on
  darwin/BSD, where the constant does not exist.

cmd/ps3netsrv-go/server.go now calls osutil.NewStrictSystemRoot.
